trabalhando nos pacotes controller e repository

+++ trabalho1/src/main/java/br/edu/iftm/tspi/pmvc/trabalho1/domain/Disciplina.java
@@ -24,7 +24,7 @@ public class Disciplina {
         this.aluno = aluno;
     }
 
-    public int getCodigoDisciplina() {
+    public long getCodigoDisciplina() {
         return codigoDisciplina;
     }
